Put everything in the correct folders
