+++ css/names.css
@@ -1,0 +1,67 @@
+canvas {
+    position:relative;
+    left:50%;
+
+    transform:translateX(-50%);
+    -webkit-transform:translateX(-50%);
+}
+
+#placeNameDiv {
+    display:block;
+    width:100%;
+    
+    margin-bottom:3%;
+}
+
+#placeNameDiv img, #placeNameDiv h1, #placeNameDiv p {
+    width:80%;
+    margin:0% 10%;
+    padding:0%;
+}
+
+#placeNameDiv img {
+    margin-bottom:2%;
+}
+
+#placeNameDiv h1, #placeNameDiv p {
+    text-align:center;
+    font-family:Helvetica;
+}
+
+#placeNameDiv h1 {
+    color:#6094c5;
+}
+
+#placeNameDiv p {
+    margin-top:2%;
+    margin-bottom:2%;
+
+    color:#555555;
+    font-size:18px;
+}
+
+#placeNameDiv a {
+    display:inline-block;
+
+    position:relative;
+    left:50%;
+
+    transform:translateX(-50%);
+    -webkit-transform:translateX(-50%);
+
+    margin:0.5% 0%;
+    padding:0.5% 1%;
+
+    color:#669ab9;
+    border:2px solid #669ab9;
+    border-radius:10px;
+
+    text-decoration:none;
+    font-family:Helvetica;
+    font-size:20px;
+}
+
+#placeNameDiv a:hover {
+    color:#336786;
+    background-color:#ace0ff;
+}
