--- placenames/index.php
@@ -1,20 +1,0 @@
-<!DOCTYPE html>
-
-<html>
-	<head>
-		<title>Place Names</title>
-		<link rel='stylesheet' type='text/css' href='placeNames.css'>
-		<script src='placeNames.js'></script>
-		<script src='profanityFilter.js'></script>
-	<head>
-	<body>
-		<canvas id='placeNameCloud' width='800' height='500'></canvas>
-		<img src='placeNames.png' alt="Image describing Place Names">
-
-		<h1>Place Names are the .com's & .net's of the Metaverse</h1>
-		<p>Where domain names get you the web pages, Place Names get you to a fixed location in 3D space.</p>
-		<p>Entering the Place Name 'Japan' might bring you to a Welcome Area somewhere in Tokyo. The exact location for the Place Name is determined by its owner.</p>
-		<p>They cost $20/year</p>
-		<a href=''>Buy a Place Name</a>
-	</body>
-</html>
