--- placenames/placeNames.css
@@ -1,64 +1,0 @@
-html, body {
-	width:100%;
-	height:100%;
-
-	margin:0%;
-	padding:0%;
-}
-
-canvas {
-	position:relative;
-	left:50%;
-
-	transform:translateX(-50%);
-	-webkit-transform:translateX(-50%);
-}
-
-img, h1, p {
-	width:60%;
-	margin:0% 20%;
-	padding:0%;
-}
-
-h1, p {
-	text-align:center;
-	font-family:Helvetica;
-}
-
-h1 {
-	color:#6094c5;
-}
-
-p {
-	margin-top:2%;
-	margin-bottom:2%;
-
-	color:#555555;
-	font-size:18px;
-}
-
-a {
-	display:block;
-	float:left;
-
-	position:relative;
-	left:50%;
-
-	transform:translateX(-50%);
-	-webkit-transform:translateX(-50%);
-
-	margin:0.5% 0%;
-	padding:0.5% 1%;
-
-	color:#669ab9;
-	border:2px solid #669ab9;
-	border-radius:10px;
-
-	text-decoration:none;
-	font-family:Helvetica;
-	font-size:20px;
-}
-
-a:hover {
-	background-color:#eeeeff;
-}
